Include mechanism for leveling out connection speeds.
Do not issue reads unless the different data channels within a stripe
are within a certain offset of each other: this implies the memory
needed to reassemble the multiple streams into a single one can be
capped.

The most difficult part is the heuristic-based determination of the
block size in mode E (without knowing the block size, when we stop
reading, the client may stop sending indefinitely).  This is done
by looking at the initial offsets of each connection, then doing
the pairwise minimum difference.

This was built on a prior series of patches, also by Brian Bockelman